Moved ContainerNodeDescriptor to its own file
