+++ lluvia/cpp/core/include/lluvia/core.h
@@ -31,18 +31,18 @@
 
 #include "core/node/ComputeNode.h"
 #include "core/node/ComputeNodeDescriptor.h"
+#include "core/node/ContainerNodeDescriptor.h"
 #include "core/node/Node.h"
+#include "core/node/NodeBuilderDescriptor.h"
 #include "core/node/NodeState.h"
 #include "core/node/NodeType.h"
 #include "core/node/PortDescriptor.h"
 #include "core/node/PortDirection.h"
 #include "core/node/PortType.h"
 #include "core/node/PushConstants.h"
-#include "core/node/NodeBuilderDescriptor.h"
 
 #include "core/CommandBuffer.h"
 #include "core/ContainerNode.h"
-#include "core/ContainerNodeDescriptor.h"
 #include "core/Duration.h"
 #include "core/error.h"
 #include "core/Interpreter.h"
