+++ lluvia/cpp/core/src/Interpreter.cpp
@@ -11,7 +11,6 @@
 #include "lluvia/core/CommandBuffer.h"
 #include "lluvia/core/ComputeDimension.h"
 #include "lluvia/core/ContainerNode.h"
-#include "lluvia/core/ContainerNodeDescriptor.h"
 
 #include "lluvia/core/image/Image.h"
 #include "lluvia/core/image/ImageAddressMode.h"
@@ -29,6 +28,7 @@
 
 #include "lluvia/core/node/ComputeNode.h"
 #include "lluvia/core/node/ComputeNodeDescriptor.h"
+#include "lluvia/core/node/ContainerNodeDescriptor.h"
 #include "lluvia/core/node/Node.h"
 #include "lluvia/core/node/NodeBuilderDescriptor.h"
 #include "lluvia/core/node/PushConstants.h"
