+++ lluvia/cpp/core/src/Session.cpp
@@ -13,8 +13,8 @@
 #include "lluvia/core/CommandBuffer.h"
 #include "lluvia/core/node/ComputeNode.h"
 #include "lluvia/core/node/ComputeNodeDescriptor.h"
+#include "lluvia/core/node/ContainerNodeDescriptor.h"
 #include "lluvia/core/ContainerNode.h"
-#include "lluvia/core/ContainerNodeDescriptor.h"
 #include "lluvia/core/Duration.h"
 #include "lluvia/core/error.h"
 #include "lluvia/core/image/Image.h"
