+++ lluvia/python/src/lluvia/core/node_new/__init__.py
@@ -8,6 +8,7 @@
 
 from lluvia.core.node_new.compute_node import *
 from lluvia.core.node_new.compute_node_descriptor import *
+from lluvia.core.node_new.container_node_descriptor import *
 from lluvia.core.node_new.node_builder_descriptor import *
 from lluvia.core.node_new.node_state import *
 from lluvia.core.node_new.node_type import *
